+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Types of changes include:
 ## Upcoming/Unreleased
 
 - CHANGED: The Enter Progress form will now default to the last measure (words, time, etc.) used when entering progress.
+- CHANGED: The copy on the homepage has been improved. The screenshots now also change with color scheme and device size.
 - FIXED: Navigating using the sidebar on mobile will now once again close the sidebar after navigation.
 - FIXED: Stat tiles now look a bit better on mobile. No more accidental masonry effect!